refactor(item): decompose the product page and keep outlet context
pages/item/[slug].tsx was 863 lines holding server props, three SWR calls,
seven useState, four async handlers, a duplicate GameApi type, 50 lines of
markdown CSS and three modals in one component. It is now ~80 lines of
routing over useItemController plus a view under
components/storefront/default/item/.

The markdown stylesheet moves to app/global.css. It was re-emitted per
render from a <style jsx global> block despite being identical every time.

Outlet context now survives the click, which it did not before:

- getServerSideProps resolves ?store= server-side and passes the outlet
  down, so the nav keeps the outlet's identity instead of reverting to
  generic Manifold, and "Back" returns to the outlet rather than /store.
  A bad ?store= degrades to an unattributed page rather than 404ing a game
  that exists, since the value is visitor-supplied.
- StoreTopNav's search autocomplete linked to /item/<slug> with no ?store=
  at all, and "View all results" always went to the global /search. A
  visitor who searched from an outlet's own header lost attribution
  entirely. Both now carry the outlet.
- The item page's own server fetch forwards x-vercel-ip-country, so a
  Brazilian visitor no longer sees BRL in the list and USD on the detail
  page for the same game.

Two hardcoded values are preserved as-is with TODOs, because changing them
changes what customers see and that does not belong in a refactor: isDemo
is pinned true, and the three feature MetaTags ignore game.meta_tags.
